Enable user to chain more than two numbers
I removed the getVariables function and variableA,variableB and
operator. I then added to getSolution() a variable that will
hold the displayValue as an array that is split to numbers and
operators. After that I added a loop, which calls on operate()
each time on the first three items of displayArr, and then
changes displayArr making it ready for the next iteration.
I updated the content of display section to be displayArr
which is the final result.