Fix bugs that rose from last commit
After the last commit I noticed that if I press clear, I later
can't add an operator to 0. Also I noticed that I can add an
operator even tough the previous onr was also an operator, if
I enter a number and delete it and then enter an operator.
To fix these, I added a function that will reset the previous
char in the display (the previous operator).